Replace FileVoque with FileSystemNodeVoque
