+++ packages/voictents-and-estinants-engine/src/custom/programmable-units/file/file.ts
@@ -1,5 +1,17 @@
+import { Merge } from 'type-fest';
+import {
+  ObjectWithPrototype,
+  buildConstructorFunctionWithName,
+} from '../../../utilities/buildConstructorFunction';
 import { FileExtensionSuffixIdentifier } from './fileExtensionSuffixIdentifier';
-import { InMemoryOdeshin2ListVoque } from '../../../core/engine/inMemoryOdeshinVoictent2';
+import {
+  BaseFileSystemNode,
+  FileSystemNodePrototype,
+  getFileSystemNodePathPartList,
+  getFileSystemNodeZorn,
+} from './fileSystemNode';
+import { FileSystemNodeVoque } from './fileSystemNodeVoictent';
+import { SpreadN } from '../../../utilities/spreadN';
 
 export type FileName = {
   pascalCase: string;
@@ -8,29 +20,57 @@ export type FileName = {
   kebabCase: string;
 };
 
-/**
- * Represents a file system file and encapuslates some naming conventions
- */
-export type File<
-  TFileExtensionSuffixIdentifier extends FileExtensionSuffixIdentifier = FileExtensionSuffixIdentifier,
-  TAdditionalMetadata extends object | null = null,
+type File2ExtensionMetadata<
+  TFileExtensionSuffixIdentifier extends FileExtensionSuffixIdentifier,
 > = {
-  zorn: string;
-  instanceId: string;
-  filePath: string;
-  directoryPath: string;
-  onDiskFileName: FileName;
-  inMemoryFileName: FileName;
-  extension: {
-    parts: string[];
-    suffix: string;
-    suffixIdentifier: TFileExtensionSuffixIdentifier;
-  };
-  additionalMetadata: TAdditionalMetadata;
+  /** @deprecated in favor of partList */
+  parts: string[];
+  partList: string[];
+  suffix: string;
+  suffixIdentifier: TFileExtensionSuffixIdentifier;
 };
 
-export const FILE_GEPP = 'file';
+type BaseFile2<
+  TFileExtensionSuffixIdentifier extends FileExtensionSuffixIdentifier = FileExtensionSuffixIdentifier,
+  TAdditionalMetadata extends object | null = null,
+> = SpreadN<
+  [
+    BaseFileSystemNode,
+    {
+      /** @deprecated in favor of zorn.forMachine  */
+      instanceId: string;
+      directoryPath: string;
+      onDiskFileName: Merge<FileName, { asIs: string }>;
+      inMemoryFileName: FileName;
+      extension: File2ExtensionMetadata<TFileExtensionSuffixIdentifier>;
+      additionalMetadata: TAdditionalMetadata;
+    },
+  ]
+>;
 
-type FileGepp = typeof FILE_GEPP;
+// TODO: move extension logic to a getter
+type File2Prototype = FileSystemNodePrototype;
 
-export type FileVoque = InMemoryOdeshin2ListVoque<FileGepp, File>;
+/**
+ * Represents a file system file
+ */
+export type File2<
+  TFileExtensionSuffixIdentifier extends FileExtensionSuffixIdentifier = FileExtensionSuffixIdentifier,
+  TAdditionalMetadata extends object | null = null,
+> = ObjectWithPrototype<
+  BaseFile2<TFileExtensionSuffixIdentifier, TAdditionalMetadata>,
+  File2Prototype
+>;
+
+export const { File2Instance } = buildConstructorFunctionWithName(
+  'File2Instance',
+)<BaseFile2, File2Prototype, File2>({
+  zorn: getFileSystemNodeZorn,
+  filePathPartList: getFileSystemNodePathPartList,
+});
+
+export const FILE_2_GEPP = 'file-2';
+
+type File2Gepp = typeof FILE_2_GEPP;
+
+export type File2Voque = FileSystemNodeVoque<File2Gepp, File2>;
